daemon: Don't unlink the pidfile if it's NULL
Bad Things will happen when we access the null pointer.

+++ daemonize.c
@@ -21,7 +21,8 @@ static const char *daemon_pidfile;
  */
 void daemon_shutdown(void)
 {
-	unlink(daemon_pidfile);
+	if (daemon_pidfile)
+		unlink(daemon_pidfile);
 }
